Move MetricNames definition into shared tracing constants.py file
